backend/requirements.txt
fastapi>=0.115,<0.116
uvicorn[standard]>=0.30,<0.31
oracledb>=2.5,<3
pandas>=2.2,<3
pydantic-settings>=2.6,<3
